Melhorando a calculadora

+++ src/br/com/dio/calculadora/Calculadora.java
@@ -6,17 +6,17 @@ public class Calculadora {
 	public static void main(String[] args) {
 		Scanner scan = new Scanner(System.in);
 		
-		int a, b;
+		double a, b;
 		
 		System.out.println("Digite o primeiro valor: ");
-		a = scan.nextInt();
+		a = scan.nextDouble();
 		System.out.println("Digite o segundo valor: ");
-		b = scan.nextInt();
+		b = scan.nextDouble();
 		
-		int soma = soma(a,b);
-		int subtracao = subtracao(a,b);
-		int divisao = divisao(a,b);
-		int multiplicacao = multiplicacao(a,b);
+		double soma = soma(a,b);
+		double subtracao = subtracao(a,b);
+		double divisao = divisao(a,b);
+		double multiplicacao = multiplicacao(a,b);
 		
 		System.out.println("soma: "+soma);
 		System.out.println("subtração: "+subtracao);
@@ -24,19 +24,19 @@ public class Calculadora {
 		System.out.println("multiplicação: "+multiplicacao);
 	}
 	
-	public static int soma(int a, int b) {
+	public static double soma(double a, double b) {
 		return a + b;
 	}
 	
-	public static int subtracao(int a, int b) {
+	public static double subtracao(double a, double b) {
 		return a - b;
 	}
 	
-	public static int divisao(int a, int b) {
+	public static double divisao(double a, double b) {
 		return a / b;
 	}
 	
-	public static int multiplicacao(int a, int b) {
+	public static double multiplicacao(double a, double b) {
 		return a * b;
 	}
 }
